281: move interpret classes into interpret/class/
