+++ src/interpret/class/Polypoint.js
@@ -1,4 +1,4 @@
-import { Point } from './point';
+import { Point } from './Point';
 
 
 /**
