+++ src/interpret/test/mapToGraph.spec.js
@@ -3,7 +3,7 @@ import test from 'tape';
 import { PPTL } from '../../global/constants';
 import { setup, teardown } from './test.spec';
 import { setupTiles, teardownTiles } from './setupTiles.spec';
-import { Point } from '../point';
+import { Point } from '../class/Point';
 import {
   mapToEdgeTiles,
   graphFromTagproMap,
